Moved parse_course_catalogue to escape testing scope
