+++ amherst_coursework_backend/amherst_coursework_algo/parse_course_catalogue/parse_course_catalogue.py
